--- Lombiq.Hosting.Tenants.Maintenance/Maintenance/DeleteMediaCache/DeleteMediaCacheMaintenanceOptions.cs
@@ -1,6 +1,0 @@
-namespace Lombiq.Hosting.Tenants.Maintenance.Maintenance.DeleteMediaCache;
-
-public class DeleteMediaCacheMaintenanceOptions
-{
-    public bool IsEnabled { get; set; }
-}
