+++ Lombiq.Hosting.Tenants.Maintenance/Maintenance/PurgeMediaCache/PurgeMediaCacheMaintenanceOptions.cs
@@ -1,0 +1,6 @@
+namespace Lombiq.Hosting.Tenants.Maintenance.Maintenance.PurgeMediaCache;
+
+public class PurgeMediaCacheMaintenanceOptions
+{
+    public bool IsEnabled { get; set; }
+}
